+++ crates/cubecl-linalg/src/matmul/components/global/load/loader/sync_buffer_loader.rs
@@ -5,9 +5,9 @@ use super::TaskCounter;
 use crate::matmul::components::InputIdent;
 use crate::matmul::components::MatmulPrecision;
 use crate::matmul::components::global::GlobalConfig;
-use crate::matmul::components::global::LoadingValidation;
 use crate::matmul::components::global::Quantization;
 use crate::matmul::components::global::load::LoadingJob;
+use crate::matmul::components::global::load::LoadingValidation;
 use crate::matmul::components::global::tensor_view::TensorReader;
 use crate::matmul::components::stage::BufferReader;
 use crate::matmul::components::stage::StageMemory;
