+++ crates/cubecl-linalg/src/matmul/components/global/load/strategy/sync_full_cyclic_checked.rs
@@ -3,14 +3,14 @@ use std::marker::PhantomData;
 use crate::matmul::components::MatmulPrecision;
 use crate::matmul::components::global::load::{SyncFullLoadingStrategy, sync_full_cyclic};
 use crate::matmul::components::global::tensor_view::TensorReader;
-use crate::matmul::components::global::{GlobalConfig, LoadingValidation, Quantization};
+use crate::matmul::components::global::{GlobalConfig, Quantization};
 use crate::matmul::components::stage::{ContiguousTilingLayout, StageMemory, TilingOrder};
 use crate::matmul::components::{Ident, InputIdent, InvalidConfigError};
 use cubecl_core as cubecl;
 use cubecl_core::prelude::*;
 use cubecl_std::CubeOption;
 
-use super::LoadingJob;
+use super::{LoadingJob, LoadingValidation};
 
 #[derive(CubeType, Clone, Copy)]
 /// Same as sync_full_cyclic but the guard is runtime
