+++ crates/cubecl-linalg/src/matmul/components/global/load/strategy/sync_full_strided.rs
@@ -1,14 +1,14 @@
 use crate::matmul::components::MatmulPrecision;
 use crate::matmul::components::global::load::SyncFullLoadingStrategy;
 use crate::matmul::components::global::tensor_view::TensorReader;
-use crate::matmul::components::global::{GlobalConfig, LoadingValidation, Quantization};
+use crate::matmul::components::global::{GlobalConfig, Quantization};
 use crate::matmul::components::stage::{StageMemory, StridedTilingLayout};
 use crate::matmul::components::{Ident, InputIdent, InvalidConfigError};
 use cubecl_core as cubecl;
 use cubecl_core::prelude::*;
 use cubecl_std::{CubeOption, CubeOptionExpand};
 
-use super::LoadingJob;
+use super::{LoadingJob, LoadingValidation};
 
 #[derive(CubeType, Clone, Copy)]
 /// Loads the content of all the tensor view using all planes,
